\name{IVAS-deprecated}
\alias{IVAS-deprecated}
\title{Deprecated functions in package \sQuote{IVAS}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{IVAS} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{MsqtlFinder: \code{\link{sQTLsFinder}}}
    \item{sqtlfinder: \code{\link{sQTLsFinder}}}
    \item{calSignificant: \code{\link{Splicingfinder}}}

  }
}
