\name{IVAS-package}
\alias{IVAS-package}
\alias{IVAS}
\docType{package}
\title{IVAS : Identification of genomic variants affecting Alternative Splicing}
\description{The tool is to detect genomic variants affecting the alternative splicing using genotypic and gene expression data(RNA-seq).}
\keyword{package}
