\name{sampleexp}
\alias{sampleexp}
\docType{data}
\title{
CEU expression data
}
\description{
CEU expression data including 78 individuals
}
\usage{data("sampleexp")}
\format{
  A data frame with 108 transcript expressions on the 78 individuals
}
\value{
  A data frame with 108 transcript expressions on the 78 individuals
}
\source{
The data was generated by GEUVADIS (Genetic European Variation in Health and Disease, A European Medical Sequencing Consortium) RNA sequencing project for 1000 Genomes samples (http://www.geuvadis.org/web/geuvadis/RNAseq-project).
}
\references{
Tuuli Lappalainen, et al. (2013). Transcriptome and genome sequencing uncovers functional variation in humans. Nature 501, 506-511.
}
\examples{
data(sampleexp)
}
\keyword{datasets}