\name{samplesnp}
\alias{samplesnp}
\docType{data}
\title{
  CEU genotype data
  }
\description{
  CEU genotype data including 78 individuals
  }
\usage{data("samplesnp")}
\format{
  A data frame with 11 SNPs on the 78 individuals
  }
\value{
  A data frame with 11 SNPs on the 78 individuals
  }
\source{
  The data has 1000 genomes Phages 1 dataset and was imputed by GEUVADIS (Genetic European Variation in Health and Disease, A European Medical Sequencing Consortium) RNA sequencing project for 1000 Genomes samples (http://www.geuvadis.org/web/geuvadis/RNAseq-project).
  }
\references{
  Tuuli Lappalainen, et al. (2013). Transcriptome and genome sequencing uncovers functional variation in humans. Nature 501, 506-511.
  }
\examples{
  data(samplesnp)
  }

\keyword{datasets}