\name{samplesnplocus}
\alias{samplesnplocus}
\docType{data}
\title{
snplocus
}
\description{
snplocus
}
\usage{data("samplesnplocus")}
\format{
  A data frame with 11 SNPs and locus of them
}
\value{
  A data frame with 11 SNPs and locus of them
}
\examples{
data(samplesnplocus)
}
\keyword{datasets}
